Compile fixes to allow compiling with VC++ 6
git-svn-id: file:///local/rexx/oorexx/oorexx.svn@3125 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ incubator/RexxGTK/4x/grxliststore.cpp
@@ -139,8 +139,8 @@ RexxMethod3(int,                       // Return type
     size_t members = context->ArraySize(args);
     int i, col, ival;
     unsigned int uival;
-    long long ival64;
-    unsigned long long uival64;
+    int64_t ival64;
+    uint64_t uival64;
     float fval;
     double dval;
     char *val;
